refactor: Consolidate generatedFilesQuestion code
